+++ lib/active_record/connection_adapters/sqlserver/type/data.rb
@@ -27,6 +27,12 @@ module ActiveRecord
           end
 
           def eql?(other)
+            # Support comparing `Type::Char`, `Type::Varchar` and `VarcharMax` with strings.
+            # This happens when we use enum with string columns.
+            if other.is_a?(::String)
+              return type.is_a?(ActiveRecord::ConnectionAdapters::SQLServer::Type::String) && value == other
+            end
+
             self.class == other.class && value == other.value
           end
           alias :== :eql?